feat(examples): add coagent todo application
This is built off the original quickstart's todo application which
can be found below.

https://github.com/CopilotKit/example-todos-app

It expands on it by adding in CoAgent support. The agent it uses
is able to manage todos and search the internet. As it does this
the state of the LangGraph agent is represented in the UI.

This works both ways, editing the todos in the UI will update them
in the Agent's state.

Not included with this is a hosted version or shot-out in the README.md,
if those are desired they will be added in a follow-up.

Signed-off-by: Tyler Slaton <tyler@copilotkit.ai>