feat(examples): add new coagents-travel example
This adds a new example in the form of a travel agent. Its operations are pretty simple but do
good job of introducing users to what can be accomplished with agentic experiences. The most
interersting part of this is combining the Agent with the ability to use the google places API
to get concrete locations to visit.

Alongside this is all of the features of CoAgents that would expect.
- Human in the loop
- Intermediate State Rendering
- Shared State
- Front-end functions

I've included some documentation in critical areas relevant to CoAgents, mostly in the
ui/lib/hooks directory.

Signed-off-by: Tyler Slaton <tyler@copilotkit.ai>